Battling Scene + Satchel UI + Databases (CouCou, Abilities)
Complete rework of how scriptable objects will be used.
Satchel nearing completion. It's only missing the item/CouCou models + scripts for that and being able to use the "Use Item" or "Change CouCou" button.
Will be commited to main to remove the branch and create a new one since this branch was redundant a while ago.