Updated Databases, Improved Interactions and Satchel, Added Element Icons, Added Elemental Advantage and created testing world.
Each CouCou now has it's own set of abilities to use.
Interactions are easier to use with new public variables to determine what it is.
Fixed bugs with the Satchel not displaying correctly and displaying buttons it shouldn't. I also added the changing of slots for the CouCou party.
Element icons have been added so the user knows which element the CouCou is.
Element advantage has also been added to the battling.
The testing scene is being built into a scene I can build and ship for testing. This includes interactions with items, picking a starter CouCou and letting the player battle with a wild CouCou.
The options menu has been given UI to script with.