Started adding fishing, updated databases, improved the stun mechanic and bugfixes
Fishing has been started. The collision with the edge of the world shows the fishing prompt. You will get correctly placed in the world according to where you collided with the edge.

Buffed some abilities as they seemed useless in battle.

Stunning can now be escaped from with rng. This is calculated as an integer from 0 to 100 - CouCou's current mindset.

Bugfixes:
- The dialogue text and speaker text now correctly hide themselves when the speaker has nothing to say