Save Game, Spawn CouCou, Move NPC, Started Audio
I created the prefab for the mailbox and added the save function to it. It saves only the players inventory and data.
NPC movement is done with multiple transforms in a list. The NPC first turns towards the transform and then moves towards it.
Spawning CouCou is done by checking when the player walks into an area, choosing a random number between 1 and 15 and waiting that number of seconds. After that it will force the player into battle.
Audio has been started with the addition of a very basic manager and triggers to put on colliders when the map is done.