Downloaded music, added credits, created a door and fixed fishing.
I've added the music that Travis made into the game and added credits to the sampler he used in order to not have to worry about the legal side of things.
I created a door with appropriate open and close animations when the player walks up to it.
Fishing seemed to easy, so I toned down how long a fish will remain on the lure from 3 seconds to 1.8 seconds.