Optimized the world + Completed boundaries and spawners + Added Quest 1
I optimized the world so that it no longer runs on 100% cpu. This required me to change lots of objects to static, materials to use gpu instancing and have the water a still colour rather than letting it reflect.
The edges of the world were added so the player can fish at their location. The player can also walk through the tall grass and expect to find a CouCou sooner or later.
Completed quest 1, full with interactions, boundaries, triggers and animations. I've started work on quest 2 as the player warps from the CouCorp to his house through a transition.