Fixing up the optimization to make the world look better
I also removed the occlusion culling as it didn't have any affect on the performance of the game.
Lighting became an issue as scenes became completely dark upon loading them in the build. This was fixed by creating a lighting settings and unchecking baked global illuminations and generating the light.
There was also a few errors on the title screen because of a few scripts being changed elsewhere. These have all been fixed up.